creates InputSection component for page 2, adds background, color picker, scheme buttons
