+++ src/components/InputForm.css
@@ -1,7 +1,7 @@
 .scheme-container {
   width: 434px;
   height: 114px;
-  background: #f1ebe5;
+  /* background: #f1ebe5; */
   border-radius: 20px;
 }
 .scheme-buttons-container {
